requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings

httpx

PyMuPDF
python-doctr[torch]
opencv-python
numpy

azure-ai-formrecognizer
azure-core

python-dotenv

